+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 11.04.2018
+
+* [Rust] Integer Complexity 1 - Prime factor test (Reddit)
+
 #### 10.04.2018
 
 * [Rust] Integer Complexity 1 - Prime numbers (Reddit)